+++ src/core/modules/NomenclatureItem.cpp
@@ -52,7 +52,7 @@ NomenclatureItem::NomenclatureItem(PlanetP nPlanet,
 	, longitude(nLongitude)
 	, size(nSize)
 {
-	StelUtils::spheToRect((longitude /*+ planet->getAxisRotation()*/) * M_PI/180.0, latitude * M_PI/180.0, XYZpc);
+	StelUtils::spheToRect((static_cast<double>(longitude) /*+ planet->getAxisRotation()*/) * M_PI/180.0, static_cast<double>(latitude) * M_PI/180.0, XYZpc);
 }
 
 NomenclatureItem::~NomenclatureItem()
@@ -736,7 +736,7 @@ float NomenclatureItem::getSelectPriority(const StelCore* core) const
 		// The planet is too faint for view (in the deep shadow for example), so let's disable select the nomenclature
 		return StelObject::getSelectPriority(core)+25.f;
 	}
-	else if (getFlagLabels() && (getAngularSize(core)*M_PI/180.*core->getProjection(StelCore::FrameJ2000)->getPixelPerRadAtCenter()>=25.))
+	else if (getFlagLabels() && (getAngularSize(core)*M_PI/180.*static_cast<double>(core->getProjection(StelCore::FrameJ2000)->getPixelPerRadAtCenter())>=25.))
 	{
 		// The item may be good selectable when it over 25px size only
 		return StelObject::getSelectPriority(core)-25.f;
@@ -827,7 +827,7 @@ Vec3d NomenclatureItem::getJ2000EquatorialPos(const StelCore* core) const
 	jde = core->getJDE();
 	const Vec3d equPos = planet->getJ2000EquatorialPos(core);
 	// Calculate the radius of the planet. It is necessary to re-scale it
-	const double r = planet->getEquatorialRadius() * planet->getSphereScale();
+	const double r = planet->getEquatorialRadius() * static_cast<double>(planet->getSphereScale());
 
 	Vec3d XYZ0;
 //	// For now, assume spherical planets, simply scale by radius.
@@ -840,7 +840,7 @@ Vec3d NomenclatureItem::getJ2000EquatorialPos(const StelCore* core) const
 	   planet->getRotEquatorialToVsop87() gives us the rotation matrix between Equatorial (on date) coordinates and Ecliptic J2000 coordinates.
 	   So we have to make another change to obtain the rotation matrix using Equatorial J2000: we have to multiplay by core->matVsop87ToJ2000 */
 	// TODO: Maybe it is more efficient to add some getRotEquatorialToVsop87Zrotation() to the Planet class which returns a Mat4d computed in Planet::computeTransMatrix().
-	XYZ = equPos + (core->matVsop87ToJ2000 * planet->getRotEquatorialToVsop87()) * Mat4d::zrotation(planet->getAxisRotation()* M_PI/180.0) * XYZ0;
+	XYZ = equPos + (core->matVsop87ToJ2000 * planet->getRotEquatorialToVsop87()) * Mat4d::zrotation(static_cast<double>(planet->getAxisRotation())* M_PI/180.0) * XYZ0;
 	return XYZ;
 }
 
@@ -852,7 +852,7 @@ float NomenclatureItem::getVMagnitude(const StelCore* core) const
 
 double NomenclatureItem::getAngularSize(const StelCore* core) const
 {
-	return std::atan2(size*planet->getSphereScale()/AU, getJ2000EquatorialPos(core).length()) * 180./M_PI;
+	return std::atan2(static_cast<double>(size*planet->getSphereScale())/AU, getJ2000EquatorialPos(core).length()) * 180./M_PI;
 }
 
 void NomenclatureItem::update(double deltaTime)
@@ -879,7 +879,7 @@ void NomenclatureItem::draw(StelCore* core, StelPainter *painter)
 			return;
 	}
 
-	double screenSize = getAngularSize(core)*M_PI/180.*painter->getProjector()->getPixelPerRadAtCenter();
+	double screenSize = getAngularSize(core)*M_PI/180.*static_cast<double>(painter->getProjector()->getPixelPerRadAtCenter());
 
 	// We can use ratio of angular size to the FOV to checking visibility of features also!
 	// double scale = getAngularSize(core)/painter->getProjector()->getFov();
@@ -890,7 +890,7 @@ void NomenclatureItem::draw(StelCore* core, StelPainter *painter)
 	if (painter->getProjector()->projectCheck(XYZ, srcPos) && (equPos.length() >= XYZ.length()) && (screenSize>50. && screenSize<750.))
 	{
 		painter->setColor(color[0], color[1], color[2], 1.0);
-		painter->drawCircle(srcPos[0], srcPos[1], 2.f);
-		painter->drawText(srcPos[0], srcPos[1], nameI18n, 0, 5.f, 5.f, false);
+		painter->drawCircle(static_cast<float>(srcPos[0]), static_cast<float>(srcPos[1]), 2.f);
+		painter->drawText(static_cast<float>(srcPos[0]), static_cast<float>(srcPos[1]), nameI18n, 0, 5.f, 5.f, false);
 	}
 }
